tutorial/pocketflow-chat/pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "Multi-provider LLM chat interface with PocketFlow"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.5.1",
    "openai>=1.3.0",
    "anthropic>=0.34.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]
